move fs output etc to fiat_shamir.rs
